don't keep separate argv config file
