+++ packages/jest-types/src/index.ts
@@ -5,7 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as Argv from './Argv';
 import * as Config from './Config';
 import * as Console from './Console';
 import * as Matchers from './Matchers';
@@ -13,4 +12,4 @@ import * as SourceMaps from './SourceMaps';
 import * as TestResult from './TestResult';
 import * as Global from './Global';
 
-export {Argv, Config, Console, Matchers, SourceMaps, TestResult, Global};
+export {Config, Console, Matchers, SourceMaps, TestResult, Global};
